[32기 정덕우] 메인페이지 구현완료
