--- src/components/ItemList/itemList.scss
@@ -1,5 +1,0 @@
-.limitBox {
-  display: flex;
-  flex-direction: row-reverse;
-  width: 100%;
-}
